Reorganise API requests and response classes
They were previously breaking the two-class-per-file rule.
