+++ osu.Game.Tests/Visual/TestCaseBeatmapScoresContainer.cs
@@ -6,7 +6,6 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.MathUtils;
 using osu.Game.Graphics;
-using osu.Game.Online.API.Requests;
 using osu.Game.Overlays.BeatmapSet.Scores;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Rulesets.Osu.Mods;
@@ -15,6 +14,7 @@ using osu.Game.Users;
 using System.Collections.Generic;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Beatmaps;
+using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Rulesets.Osu;
 
 namespace osu.Game.Tests.Visual
@@ -22,9 +22,9 @@ namespace osu.Game.Tests.Visual
     [System.ComponentModel.Description("in BeatmapOverlay")]
     public class TestCaseBeatmapScoresContainer : OsuTestCase
     {
-        private readonly IEnumerable<OnlineScore> scores;
-        private readonly IEnumerable<OnlineScore> anotherScores;
-        private readonly OnlineScore topScore;
+        private readonly IEnumerable<APIScore> scores;
+        private readonly IEnumerable<APIScore> anotherScores;
+        private readonly APIScore topScore;
         private readonly Box background;
 
         public TestCaseBeatmapScoresContainer()
@@ -57,7 +57,7 @@ namespace osu.Game.Tests.Visual
 
             scores = new[]
             {
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -80,7 +80,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 1234567890,
                     Accuracy = 1,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -102,7 +102,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 1234789,
                     Accuracy = 0.9997,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -123,7 +123,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 12345678,
                     Accuracy = 0.9854,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -143,7 +143,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 1234567,
                     Accuracy = 0.8765,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -169,7 +169,7 @@ namespace osu.Game.Tests.Visual
 
             anotherScores = new[]
             {
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -191,7 +191,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 1234789,
                     Accuracy = 0.9997,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -214,7 +214,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 1234567890,
                     Accuracy = 1,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -230,7 +230,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 123456,
                     Accuracy = 0.6543,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -251,7 +251,7 @@ namespace osu.Game.Tests.Visual
                     TotalScore = 12345678,
                     Accuracy = 0.9854,
                 },
-                new OnlineScore
+                new APIScore
                 {
                     User = new User
                     {
@@ -279,7 +279,7 @@ namespace osu.Game.Tests.Visual
                 s.Statistics.Add(HitResult.Meh, RNG.Next(2000));
             }
 
-            topScore = new OnlineScore
+            topScore = new APIScore
             {
                 User = new User
                 {
