+++ osu.Game/Online/API/Requests/GetBeatmapDetailsRequest.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using Newtonsoft.Json;
 using osu.Game.Beatmaps;
+using osu.Game.Online.API.Requests.Responses;
 
 namespace osu.Game.Online.API.Requests
 {
-    public class GetBeatmapDetailsRequest : APIRequest<GetBeatmapDetailsResponse>
+    public class GetBeatmapDetailsRequest : APIRequest<APIBeatmapMetrics>
     {
         private readonly BeatmapInfo beatmap;
 
@@ -18,29 +18,5 @@ namespace osu.Game.Online.API.Requests
         }
 
         protected override string Target => $@"beatmaps/{lookupString}";
-    }
-
-    public class GetBeatmapDetailsResponse : BeatmapMetrics
-    {
-        //the online API returns some metrics as a nested object.
-        [JsonProperty(@"failtimes")]
-        private BeatmapMetrics failTimes
-        {
-            set
-            {
-                Fails = value.Fails;
-                Retries = value.Retries;
-            }
-        }
-
-        //and other metrics in the beatmap set.
-        [JsonProperty(@"beatmapset")]
-        private BeatmapMetrics beatmapSet
-        {
-            set
-            {
-                Ratings = value.Ratings;
-            }
-        }
     }
 }
