+++ osu.Game/Online/API/Requests/GetBeatmapSetRequest.cs
@@ -1,9 +1,11 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
+using osu.Game.Online.API.Requests.Responses;
+
 namespace osu.Game.Online.API.Requests
 {
-    public class GetBeatmapSetRequest : APIRequest<APIResponseBeatmapSet>
+    public class GetBeatmapSetRequest : APIRequest<APIBeatmapSet>
     {
         private readonly int id;
         private readonly BeatmapSetLookupType type;
