+++ osu.Game/Overlays/Profile/Sections/Ranks/PaginatedScoreContainer.cs
@@ -8,6 +8,7 @@ using osu.Game.Online.API.Requests;
 using osu.Game.Users;
 using System;
 using System.Linq;
+using osu.Game.Online.API.Requests.Responses;
 
 namespace osu.Game.Overlays.Profile.Sections.Ranks
 {
@@ -49,7 +50,7 @@ namespace osu.Game.Overlays.Profile.Sections.Ranks
 
                 MissingText.Hide();
 
-                foreach (OnlineScore score in scores)
+                foreach (APIScore score in scores)
                 {
                     DrawableProfileScore drawableScore;
