+++ osu.Game/Overlays/Profile/Sections/Recent/DrawableRecentActivity.cs
@@ -8,6 +8,7 @@ using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests;
+using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Online.Chat;
 using osu.Game.Screens.Select.Leaderboards;
 
@@ -17,11 +18,11 @@ namespace osu.Game.Overlays.Profile.Sections.Recent
     {
         private APIAccess api;
 
-        private readonly RecentActivity activity;
+        private readonly APIRecentActivity activity;
 
         private LinkFlowContainer content;
 
-        public DrawableRecentActivity(RecentActivity activity)
+        public DrawableRecentActivity(APIRecentActivity activity)
         {
             this.activity = activity;
         }
